requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers==0.30.0
transformers==4.44.0
accelerate==0.33.0
peft==0.12.0
bitsandbytes==0.43.3
datasets==2.20.0
huggingface_hub==0.24.5
controlnet_aux==0.0.9

einops
Pillow
tomesd
imageio[ffmpeg]
opencv-python-headless
gradio==4.42.0
safetensors
sentencepiece
protobuf
